Tweak gazettes
