+++ lawlibrary/templates/lawlibrary/_localities.html
@@ -3,7 +3,7 @@
     <ul class="list-unstyled mb-0">
       {% for locality in group %}
         <li>
-          <a href="{% url 'provincial_gazette_list' locality.code %}">{{ locality.name }}</a>
+          <a href="{% url 'gazettes_by_locality' locality.code %}">{{ locality.name }}</a>
         </li>
       {% endfor %}
     </ul>
