+++ lawlibrary/templates/peachjam/gazette_detail.html
@@ -6,7 +6,7 @@
     <ol class="breadcrumb">
       <li class="breadcrumb-item"><a href="{% url 'gazettes' %}">{% trans 'Gazettes' %}</a></li>
       {% if document.locality %}
-        <li class="breadcrumb-item"><a href="{% url 'provincial_gazette_list' document.locality.code %}">{{ document.locality }}</a></li>
+        <li class="breadcrumb-item"><a href="{% url 'gazettes_by_locality' document.locality.code %}">{{ document.locality }}</a></li>
         <li class="breadcrumb-item"><a href="{% url 'gazettes_by_year' document.locality.code document.date.year %}">{{ document.date.year }}</a></li>
       {% else %}
         <li class="breadcrumb-item"><a href="{% url 'gazettes_by_year' year=document.date.year %}">{{ document.date.year }}</a></li>
