+++ lawlibrary/templates/peachjam/gazette_list.html
@@ -23,10 +23,32 @@
   {% endif %}
 {% endblock %}
 
-{% block gazette-year-link %}
-  {% if locality %}
-    <a class="text-decoration-none" href="{% url 'gazettes_by_year' locality.code year.year %}">{{ year.year }}</a>
-  {% else %}
-    <a class="text-decoration-none" href="{% url 'gazettes_by_year' year.year %}">{{ year.year }}</a>
-  {% endif %}
+{% block years-list %}
+  <div class="col-md-10 mt-5 mb-5">
+    <div class="row">
+      {% for year in years %}
+        <div class="col-md-4 col-sm-6 col-xs-12 status-box">
+          <div>
+            <h2>
+              {% block gazette-year-link %}
+                {% if locality %}
+                  <a class="text-decoration-none" href="{% url 'gazettes_by_year' locality.code year.year %}">{{ year.year }}</a>
+                {% else %}
+                  <a class="text-decoration-none" href="{% url 'gazettes_by_year' year.year %}">{{ year.year }}</a>
+                {% endif %}
+              {% endblock %}
+            </h2>
+            <p class="status-up"><i class="fas fa-signal" aria-hidden="true"></i>
+              {% blocktrans trimmed count num_gazettes=year.count %}
+                {{ num_gazettes }} gazette
+              {% plural %}
+                {{ num_gazettes }} gazettes
+              {% endblocktrans %}
+            </p>
+          </div>
+          <div data-year="{{ year.month }}" class="monthly-chart"></div>
+        </div>
+      {% endfor %}
+    </div>
+  </div>
 {% endblock %}
