+++ lawlibrary/templates/peachjam/gazette_year.html
@@ -6,7 +6,7 @@
     <ol class="breadcrumb">
       <li class="breadcrumb-item"><a href="{% url 'gazettes' %}">{% trans 'Gazettes' %}</a></li>
       {% if locality %}
-        <li class="breadcrumb-item"><a href="{% url 'provincial_gazette_list' code=locality.code %}">{{ locality }}</a></li>
+        <li class="breadcrumb-item"><a href="{% url 'gazettes_by_locality' code=locality.code %}">{{ locality }}</a></li>
       {% endif %}
     </ol>
   </nav>
