+++ lawlibrary/views/__init__.py
@@ -1,4 +1,3 @@
 # flake8: noqa
-from .gazette import *
 from .home import *
 from .legislation import *
