+++ peachjam/static/stylesheets/components/_year_nav.scss
@@ -1,0 +1,21 @@
+.year-nav {
+  list-style-type: none;
+  padding-left: 0px;
+
+  li {
+    display: inline;
+    margin-right: 15px;
+
+    a,
+    a:visited {
+      color: $primary;
+      text-decoration: none;
+    }
+
+    a.active,
+    a.active:visited {
+      color: black;
+      font-weight: 500;
+    }
+  }
+}
