+++ peachjam/views/gazette.py
@@ -3,11 +3,12 @@ from operator import itemgetter
 
 from django.db.models import Count
 from django.db.models.functions import ExtractMonth, ExtractYear
+from django.shortcuts import get_object_or_404
 from django.urls import reverse
 from django.utils.dates import MONTHS
 from django.views.generic import TemplateView
 
-from peachjam.models import Gazette
+from peachjam.models import Gazette, Locality
 from peachjam.registry import registry
 from peachjam.views.generic_views import BaseDocumentDetailView, DocumentListView
 
@@ -36,13 +37,32 @@ class GazetteListView(TemplateView):
     template_name = "peachjam/gazette_list.html"
     navbar_link = "gazettes"
 
+    def get(self, request, code=None, *args, **kwargs):
+        self.locality = get_object_or_404(Locality, code=code) if code else None
+        return super().get(request, *args, **kwargs)
+
     def get_queryset(self):
-        return self.queryset
+        qs = self.queryset
+        if self.locality:
+            qs = qs.filter(locality=self.locality)
+        return qs
 
     def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
+        context = super().get_context_data(locality=self.locality, **kwargs)
 
         queryset = self.get_queryset()
+
+        if self.locality is None:
+            locality_ids = list(
+                queryset.order_by()
+                .distinct("locality")
+                .values_list("locality", flat=True)
+            )
+            context["localities"] = Locality.objects.filter(pk__in=locality_ids)
+
+        provinces = list(Locality.objects.all())
+        context["province_groups"] = provinces[:5], provinces[5:]
+
         context["num_gazettes"] = queryset.count()
         context["years"] = self.get_year_stats(queryset)
 
@@ -65,6 +85,10 @@ class GazetteYearView(DocumentListView):
     navbar_link = "gazettes"
     locality = None
 
+    def get(self, request, code=None, *args, **kwargs):
+        self.locality = get_object_or_404(Locality, code=code) if code else None
+        return super().get(request, *args, **kwargs)
+
     def get_queryset(self):
         return super().get_queryset().filter(date__year=self.kwargs["year"])
 
@@ -80,6 +104,7 @@ class GazetteYearView(DocumentListView):
             .values("year", "count")
         )
         context["years"] = group_years(years, self.locality)
+        context["locality"] = self.locality
 
         context["gazettes"] = self.group_gazettes(list(self.get_queryset()))
         context["year"] = int(self.kwargs["year"])
